move test methods to their own class.
